clean up error handling of request_impl in websockets
